requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
unidecode
regex
markdown
langchain
langchain-community
langchain-ollama
langchain-chroma
chromadb
python-dotenv
aiosqlite
jinja2
python-docx
zipfile
shutil
